listener: active connection move only (#17551)
