+++ source/server/active_tcp_listener.cc
@@ -167,9 +167,9 @@ void ActiveTcpListener::newActiveConnection(const Network::FilterChain& filter_c
                                             Network::ServerConnectionPtr server_conn_ptr,
                                             std::unique_ptr<StreamInfo::StreamInfo> stream_info) {
   auto& active_connections = getOrCreateActiveConnections(filter_chain);
-  ActiveTcpConnectionPtr active_connection(
-      new ActiveTcpConnection(active_connections, std::move(server_conn_ptr),
-                              dispatcher().timeSource(), std::move(stream_info)));
+  auto active_connection =
+      std::make_unique<ActiveTcpConnection>(active_connections, std::move(server_conn_ptr),
+                                            dispatcher().timeSource(), std::move(stream_info));
   // If the connection is already closed, we can just let this connection immediately die.
   if (active_connection->connection_->state() != Network::Connection::State::Closed) {
     ENVOY_CONN_LOG(debug, "new connection from {}", *active_connection->connection_,
@@ -205,62 +205,6 @@ void ActiveTcpListener::post(Network::ConnectionSocketPtr&& socket) {
       return;
     }
   });
-}
-
-ActiveConnections::ActiveConnections(ActiveTcpListener& listener,
-                                     const Network::FilterChain& filter_chain)
-    : listener_(listener), filter_chain_(filter_chain) {}
-
-ActiveConnections::~ActiveConnections() {
-  // connections should be defer deleted already.
-  ASSERT(connections_.empty());
-}
-
-ActiveTcpConnection::ActiveTcpConnection(ActiveConnections& active_connections,
-                                         Network::ConnectionPtr&& new_connection,
-                                         TimeSource& time_source,
-                                         std::unique_ptr<StreamInfo::StreamInfo>&& stream_info)
-    : stream_info_(std::move(stream_info)), active_connections_(active_connections),
-      connection_(std::move(new_connection)),
-      conn_length_(new Stats::HistogramCompletableTimespanImpl(
-          active_connections_.listener_.stats_.downstream_cx_length_ms_, time_source)) {
-  // We just universally set no delay on connections. Theoretically we might at some point want
-  // to make this configurable.
-  connection_->noDelay(true);
-  auto& listener = active_connections_.listener_;
-  listener.stats_.downstream_cx_total_.inc();
-  listener.stats_.downstream_cx_active_.inc();
-  listener.per_worker_stats_.downstream_cx_total_.inc();
-  listener.per_worker_stats_.downstream_cx_active_.inc();
-
-  // Active connections on the handler (not listener). The per listener connections have already
-  // been incremented at this point either via the connection balancer or in the socket accept
-  // path if there is no configured balancer.
-  listener.parent_.incNumConnections();
-}
-
-ActiveTcpConnection::~ActiveTcpConnection() {
-  ActiveStreamListenerBase::emitLogs(*active_connections_.listener_.config_, *stream_info_);
-  auto& listener = active_connections_.listener_;
-  listener.stats_.downstream_cx_active_.dec();
-  listener.stats_.downstream_cx_destroy_.inc();
-  listener.per_worker_stats_.downstream_cx_active_.dec();
-  conn_length_->complete();
-
-  // Active listener connections (not handler).
-  listener.decNumConnections();
-
-  // Active handler connections (not listener).
-  listener.parent_.decNumConnections();
-}
-
-void ActiveTcpConnection::onEvent(Network::ConnectionEvent event) {
-  ENVOY_LOG(trace, "[C{}] connection on event {}", connection_->id(), static_cast<int>(event));
-  // Any event leads to destruction of the connection.
-  if (event == Network::ConnectionEvent::LocalClose ||
-      event == Network::ConnectionEvent::RemoteClose) {
-    active_connections_.listener_.removeConnection(*this);
-  }
 }
 
 } // namespace Server
