+++ source/server/active_tcp_listener.h
@@ -11,12 +11,6 @@
 
 namespace Envoy {
 namespace Server {
-
-struct ActiveTcpConnection;
-using ActiveTcpConnectionPtr = std::unique_ptr<ActiveTcpConnection>;
-class ActiveConnections;
-using ActiveConnectionCollectionPtr = std::unique_ptr<ActiveConnections>;
-
 namespace {
 // Structure used to allow a unique_ptr to be captured in a posted lambda. See below.
 struct RebalancedSocket {
@@ -106,44 +100,5 @@ public:
 };
 
 using ActiveTcpListenerOptRef = absl::optional<std::reference_wrapper<ActiveTcpListener>>;
-
-/**
- * Wrapper for a group of active connections which are attached to the same filter chain context.
- */
-class ActiveConnections : public Event::DeferredDeletable {
-public:
-  ActiveConnections(ActiveTcpListener& listener, const Network::FilterChain& filter_chain);
-  ~ActiveConnections() override;
-
-  // listener filter chain pair is the owner of the connections
-  ActiveTcpListener& listener_;
-  const Network::FilterChain& filter_chain_;
-  // Owned connections
-  std::list<ActiveTcpConnectionPtr> connections_;
-};
-
-/**
- * Wrapper for an active TCP connection owned by this handler.
- */
-struct ActiveTcpConnection : LinkedObject<ActiveTcpConnection>,
-                             public Event::DeferredDeletable,
-                             public Network::ConnectionCallbacks,
-                             Logger::Loggable<Logger::Id::conn_handler> {
-  ActiveTcpConnection(ActiveConnections& active_connections,
-                      Network::ConnectionPtr&& new_connection, TimeSource& time_system,
-                      std::unique_ptr<StreamInfo::StreamInfo>&& stream_info);
-  ~ActiveTcpConnection() override;
-
-  // Network::ConnectionCallbacks
-  void onEvent(Network::ConnectionEvent event) override;
-  void onAboveWriteBufferHighWatermark() override {}
-  void onBelowWriteBufferLowWatermark() override {}
-
-  std::unique_ptr<StreamInfo::StreamInfo> stream_info_;
-  ActiveConnections& active_connections_;
-  Network::ConnectionPtr connection_;
-  Stats::TimespanPtr conn_length_;
-};
-
 } // namespace Server
 } // namespace Envoy
